refactor: split out shared log helper into module
