+++ lib/scanner.js
@@ -1,5 +1,7 @@
 const nmap = require('node-nmap')
 
+const { log } = require('./log')
+
 const DISABLE_NMAP_DISCOVERY = '-Pn '
 const SCAN_UDP_WITH_VERSION = '-sU -sV '
 const SCAN_TIMEOUT_MILLIS = 30000
@@ -30,18 +32,18 @@ function buildNmapOptions(ports, protocol, rng = Math.random) {
 }
 
 function scan(host, ports, protocol, done) {
-  nmap.nmapLocation = '/usr/bin/nmap' // default
+  nmap.nmapLocation = '/usr/bin/nmap' // We run in a container so we can be sure of this
   const nmapOptions = buildNmapOptions(ports, protocol)
 
-  console.log(`query string: ${nmapOptions}, for ${protocol}`)
+  log(`query string: ${nmapOptions}, for ${protocol}`)
 
   let quickscan = new nmap.NmapScan(host, nmapOptions)
 
   quickscan.scanTimeout = SCAN_TIMEOUT_MILLIS
   quickscan.on('complete', scanResults => {
-    console.log(`results for ${protocol}`)
-    console.log(JSON.stringify(scanResults, null, 2))
-    console.log(`proto ${protocol}:`, JSON.stringify(ports, null, 2))
+    log(`results for ${protocol}`)
+    log(JSON.stringify(scanResults, null, 2))
+    log(`proto ${protocol}:`, JSON.stringify(ports, null, 2))
 
     return done(null, parseResults(scanResults))
   })
